fixed sorting box number issue by change int to bigint to account for large box numbers

+++ services/anchorBoxService.js
@@ -181,7 +181,7 @@ const getBoxesForSearch = async (query) => {
     res = await db.query(
       `SELECT latitude as lat, longitude as lon, box_id as display_name, zip_code, country FROM "Anchor_Box" WHERE box_id LIKE '%' || $1 || '%'
       AND latitude IS NOT NULL AND longitude IS NOT NULL AND country is NOT NULL AND zip_code IS NOT NULL
-      ORDER BY box_id::int`,
+      ORDER BY box_id::bigint`,
       [query],
     );
   } catch (err) {
